% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.interval.R
\name{get.interval}
\alias{get.interval}
\title{Get the prediction interval using a simulated data set}
\usage{
get.interval(
  METHOD = "Shortest",
  DIST = "Normal",
  n = 100,
  test_n = 100,
  alpha = 0.05,
  test_data = F,
  K = 2,
  shape = 5,
  beta,
  ...
)
}
\arguments{
\item{METHOD}{a character string or a number giving the method to be used. This must match one of "Random Position", "Shortest", "Cross Validation" or "Data Driven", with default "Shortest". 1, 2, 3, 4 refer to  "Random Position", "Shortest", "Cross Validation" and "Data Driven" respectively.}

\item{DIST}{distribution that the simulated data from. This must match one of "Normal", "Uniform", "Exponential", "Gamma", with default "Normal".}

\item{n}{Number of observations for the simulated data}

\item{test_n}{Number of observations for the test data}

\item{alpha}{nominal error rate}

\item{test_data}{whether generate the test data}

\item{K}{k-fold cross-validation. Only needed when METHOD is "Cross Validation", with default 2.}

\item{shape}{shape parameter needed for the standard Gamma distribution, with default 5.}

\item{beta}{a random position. Only needed when METHOD is "Random Position". beta must be less than or equal to alpha.}

\item{...}{arguments passed to or from distributions}
}
\value{
A list.
\item{interval}{prediction interval}
\item{data}{simulated data}
\item{test_dataset}{test data}
}
\description{
Get the prediction interval by a chosen method using a simulated data set. The simulated data are generated from standard Normal distribution, standard Uniform distribution, standard Exponential distribution or standard Gamma distribution unless more parameters for the distribution are specified.
}
\examples{
\dontrun{
# Random position method
get.interval(METHOD = "Random Position")

# Shortest method
get.interval(METHOD = "Shortest")

# Uniform distribution
get.interval(DIST = "Uniform")

# Normal distribution
get.interval(DIST = "Normal")
}

}
