% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.quantile.R
\name{lower.quantile}
\alias{lower.quantile}
\title{Calculate the lower quantile for data with weights}
\usage{
lower.quantile(v, prob, w = NULL, sorted = FALSE)
}
\arguments{
\item{v}{the data that we want to find quantile}

\item{prob}{find the lower 100*prob\% quantile}

\item{w}{weight for data}

\item{sorted}{whether v has been sorted}
}
\description{
Calculate the lower quantile for data with weights
}
\examples{
\dontrun{
example_data <- rnorm(500, 0, 5)

# Find the lower 0.975 quantile
lower.quantile(x, prob = 0.975)

# Find the lower 0.025 quantile
lower.quantile(x, prob = 0.975)
}

}
