% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.pi.R
\name{random.pi}
\alias{random.pi}
\title{Calculate prediction interval based on random position.}
\usage{
random.pi(x, alpha, beta = NULL)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}

\item{beta}{the left tail probability chosen at random. must be less than or equal to alpha.}
}
\description{
Calculate prediction interval based on random position.
}
\examples{
\dontrun{
example_data <- rnorm(500, 0, 5)
random.pi(x, alpha = 0.05, beta = 0.01)
}

}
