% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest.pi.R
\name{shortest.pi}
\alias{shortest.pi}
\title{Calculate shortest prediction interval}
\usage{
shortest.pi(x, alpha)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}
}
\description{
Calculate shortest prediction interval
}
\examples{
\dontrun{
example_data <- rnorm(500, 0, 5)
shortest.pi(example_data, alpha = 0.01)
}

}
