% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.quantile.R
\name{upper.quantile}
\alias{upper.quantile}
\title{Calculate the upper quantile for data with weights}
\usage{
upper.quantile(v, prob, sorted = FALSE)
}
\arguments{
\item{v}{the data that we want to find quantile}

\item{prob}{find the upper 100*prob\% quantile}

\item{sorted}{whether v has been sorted}
}
\description{
Calculate the upper quantile for data with weights
}
\examples{
\dontrun{
example_data <- rnorm(500, 0, 5)

# Find the upper 0.975 quantile
lower.quantile(x, prob = 0.975)

# Find the upper 0.025 quantile
lower.quantile(x, prob = 0.975)
}

}
