% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative.pi.R
\name{conservative.pi}
\alias{conservative.pi}
\title{Conservative prediction interval}
\usage{
conservative.pi(x, alpha)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}
}
\description{
Conservative prediction interval
}
