% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.quantile.R
\name{upper.quantile}
\alias{upper.quantile}
\title{Calculate the upper quantile for data with weights}
\usage{
upper.quantile(v, prob, w = NULL, sorted = FALSE)
}
\arguments{
\item{v}{the data that we want to find quantile}

\item{prob}{find the upper 100*prob\% quantile}

\item{w}{weight for data}

\item{sorted}{whether v has been sorted}
}
\description{
Calculate the upper quantile for data with weights
}
